relocate collection install-session tracking into core (483 WIP)
The install-session tracking definitions (status type, session interfaces,
reducer, actions, helpers and selectors) lived under the collections extension
but are consumed outside it by InstallManager, extendAPI, health_check and
nexus_integration. Core importing from an extension is a backwards dependency.

Move the definitions to neutral core dirs so they are globally importable
without the inversion:
- types   -> types/collections/ICollectionInstallSession.ts
- reducer -> reducers/collectionInstallTracking.ts
- actions -> actions/collectionInstallTracking.ts
- helpers + selectors -> util/collectionInstallSession.ts,
  util/collectionInstallSessionSelectors.ts

The collections extension still owns the session lifecycle and keeps
registering the reducer via context.registerReducer; only the definitions
moved. Row assembly (itemRows) stays in collections as it is UI-only.
Behavior-neutral: imports and core barrels repointed.

part of LAZ-483
