+++ src/renderer/src/extensions/collections/index.ts
@@ -15,6 +15,7 @@ import * as actions from "../../actions";
 import { installIconSet } from "../../controls/Icon";
 import OptionsFilter from "../../controls/table/OptionsFilter";
 import { log } from "../../logging";
+import trackingReducer from "../../reducers/collectionInstallTracking";
 import type { IDialogResult } from "../../types/IDialog";
 import type { IExtensionApi, IExtensionContext } from "../../types/IExtensionContext";
 import type { INotification } from "../../types/INotification";
@@ -57,7 +58,6 @@ import { onCollectionUpdate } from "./eventHandlers";
 import initIniTweaks from "./initweaks";
 import { makeInstall } from "./makeInstall";
 import { postprocessCollection } from "./postprocessCollection";
-import trackingReducer from "./reducers/installTracking";
 import persistentReducer from "./reducers/persistent";
 import sessionReducer from "./reducers/session";
 import settingsReducer from "./reducers/settings";
