+++ src/renderer/src/extensions/collections/views/CollectionPageView/index.tsx
@@ -24,10 +24,12 @@ import OptionsFilter from "../../../../controls/table/OptionsFilter";
 import TableTextFilter from "../../../../controls/table/TextFilter";
 import * as tooltip from "../../../../controls/TooltipControls";
 import { log } from "../../../../logging";
+import type { ICollectionModInstallInfo } from "../../../../types/collections/ICollectionInstallSession";
 import type { IDialogResult } from "../../../../types/IDialog";
 import type { INotification } from "../../../../types/INotification";
 import type { IOverlay, IState } from "../../../../types/IState";
 import type { ITableAttribute } from "../../../../types/ITableAttribute";
+import { getCollectionActiveSession } from "../../../../util/collectionInstallSessionSelectors";
 import { ProcessCanceled, UserCanceled } from "../../../../util/CustomErrors";
 import { showError } from "../../../../util/message";
 import { getSafe } from "../../../../util/storeHelper";
@@ -47,8 +49,6 @@ import { showUsageInstruction } from "../../../settings_interface/actions/interf
 import { AUTHOR_UNKNOWN, AVATAR_FALLBACK } from "../../constants";
 import { buildCollectionItemRows } from "../../installSession/itemRows";
 import type { ICollectionItemRow } from "../../installSession/itemRows";
-import { getCollectionActiveSession } from "../../installSession/selectors";
-import type { ICollectionModInstallInfo } from "../../installSession/types";
 import InstallDriver from "../../util/InstallDriver";
 import CollectionInstructions from "./CollectionInstructions";
 import CollectionItemStatus from "./CollectionItemStatus";
