+++ src/renderer/src/extensions/health_check/api/triggers.ts
@@ -4,8 +4,8 @@ import { log } from "../../../logging";
 import type { IExtensionApi } from "../../../types/IExtensionContext";
 import type { IHealthCheckResult } from "../../../types/IHealthCheck";
 import { HealthCheckTrigger } from "../../../types/IHealthCheck";
+import { hasCollectionActiveSession } from "../../../util/collectionInstallSessionSelectors";
 import Debouncer from "../../../util/Debouncer";
-import { hasCollectionActiveSession } from "../../collections/installSession/selectors";
 import type { IHealthCheckApi } from "../types";
 
 /**
