+++ src/renderer/src/extensions/mod_management/util/extendAPI.ts
@@ -1,9 +1,9 @@
 import type { IExtensionApi } from "../../../types/api";
-import { log } from "../../../util/log";
 import {
   getCollectionActiveSession,
   getCollectionCurrentPhase,
-} from "../../collections/installSession/selectors";
+} from "../../../util/collectionInstallSessionSelectors";
+import { log } from "../../../util/log";
 import type InstallManager from "../InstallManager";
 import type { IDeployOptions } from "../types/IDeployOptions";
 import type { IModsAPIExtension } from "../types/IModsAPIExtension";
