+++ src/renderer/src/util/api.ts
@@ -12,10 +12,6 @@ import {
   resolveCategoryName,
   resolveCategoryPath,
 } from "../extensions/category_management/util/retrieveCategoryPath";
-import {
-  generateCollectionSessionId,
-  modRuleId,
-} from "../extensions/collections/installSession/util";
 import { readExtensibleDir } from "../extensions/extension_manager/util";
 import getDriveList from "../extensions/gamemode_management/util/getDriveList";
 import { getGame, getGames } from "../extensions/gamemode_management/util/getGame";
@@ -50,6 +46,7 @@ import { getApplication } from "./application";
 import { Archive } from "./archives";
 import calculateFolderSize from "./calculateFolderSize";
 import { checksum, fileMD5 } from "./checksum";
+import { generateCollectionSessionId, modRuleId } from "./collectionInstallSession";
 import ConcurrencyLimiter from "./ConcurrencyLimiter";
 import copyRecursive from "./copyRecursive";
 import {
